+++ src/constants.ts
@@ -1,0 +1,53 @@
+// Language translations for different UI elements
+export const LANGUAGE_TRANSLATIONS: Record<string, Record<string, string>> = {
+    'en': {
+        keyDevelopments: 'Key Developments',
+        analysisContext: 'Analysis & Context',
+        processingStatus: 'Processing Status',
+        generatedAt: 'Generated at',
+        tableOfContents: 'Table of Contents',
+        noRecentNews: 'No recent news found for',
+        errorRetrieving: 'Error retrieving news for',
+        limitedNews: 'Limited substantive news found on'
+    },
+    'fr': {
+        keyDevelopments: 'Développements Clés',
+        analysisContext: 'Analyse et Contexte',
+        processingStatus: 'Statut du Traitement',
+        generatedAt: 'Généré à',
+        tableOfContents: 'Table des Matières',
+        noRecentNews: 'Aucune actualité récente trouvée pour',
+        errorRetrieving: 'Erreur lors de la récupération des actualités pour',
+        limitedNews: 'Actualités substantielles limitées trouvées sur'
+    },
+    'de': {
+        keyDevelopments: 'Wichtige Entwicklungen',
+        analysisContext: 'Analyse & Kontext',
+        processingStatus: 'Verarbeitungsstatus',
+        generatedAt: 'Erstellt am',
+        tableOfContents: 'Inhaltsverzeichnis',
+        noRecentNews: 'Keine aktuellen Nachrichten gefunden für',
+        errorRetrieving: 'Fehler beim Abrufen von Nachrichten für',
+        limitedNews: 'Begrenzte substanzielle Nachrichten gefunden zu'
+    },
+    'es': {
+        keyDevelopments: 'Desarrollos Clave',
+        analysisContext: 'Análisis y Contexto',
+        processingStatus: 'Estado del Procesamiento',
+        generatedAt: 'Generado a las',
+        tableOfContents: 'Tabla de Contenidos',
+        noRecentNews: 'No se encontraron noticias recientes para',
+        errorRetrieving: 'Error al recuperar noticias para',
+        limitedNews: 'Noticias sustanciales limitadas encontradas sobre'
+    },
+    'it': {
+        keyDevelopments: 'Sviluppi Chiave',
+        analysisContext: 'Analisi e Contesto',
+        processingStatus: 'Stato di Elaborazione',
+        generatedAt: 'Generato alle',
+        tableOfContents: 'Sommario',
+        noRecentNews: 'Nessuna notizia recente trovata per',
+        errorRetrieving: 'Errore nel recuperare notizie per',
+        limitedNews: 'Notizie sostanziali limitate trovate su'
+    }
+};
